feat: bit of refactor ... add geom to routes
